+++ libsolidity/ast/Types.cpp
@@ -3366,6 +3366,25 @@ TypeResult FunctionType::interfaceType(bool /*_inLibrary*/) const
 		return TypeResult::err("Internal type is not allowed for public or external functions.");
 }
 
+TypePointer FunctionType::mobileType() const
+{
+	// return function without parameter names
+	return TypeProvider::function(
+		m_parameterTypes,
+		m_returnParameterTypes,
+		strings(m_parameterTypes.size()),
+		m_returnParameterNames,
+		m_kind,
+		m_arbitraryParameters,
+		m_stateMutability,
+		m_declaration,
+		m_gasSet,
+		m_valueSet,
+		m_bound,
+		m_saltSet
+	);
+}
+
 bool FunctionType::canTakeArguments(
 	FuncCallArguments const& _arguments,
 	Type const* _selfType